+++ Dignus.Actor.Core/IActorRefProvider.cs
@@ -8,7 +8,7 @@ namespace Dignus.Actor.Core
 {
     internal interface IActorRefProvider
     {
-        bool TryGetActorRef(int id, out IActorRef actorRef);
+        bool TryGetActorRef(long id, out IActorRef actorRef);
 
         bool TryGetActorRef(string alias, out IActorRef actorRef);
     }
